Consolidating to a single service that just schedules a job, retrieves the details and handles job invocation to simplify the quickstart
